Fixes a bug that causes the storage of the active account to act as a singleton
It is not enough to deepcopy the storage, as all copies of that global state will still point to the same environment, pointing to the same active account, pointing to the same now overwritten active account. I noticed this because after processing a STOP, the next instruction CALLVALUE of the alternative branch had the same storage content as stop.

One solution seems to be the proposed change actually deepcopying the entire environment, but it is also necessary to ensure that when branching mstate gets a new instance.