Fix the index parameter of SymExecWrapper::find_storage_write()
SymExecWrapper uses the stringified storage indices to index the
storage.

If the storage index is a bit vector value, its __str__() implemented
in Z3 will just use its integer value in the string. And we can just
pass a stringified integer to SymExecWrapper::find_storage_write() to
get the corresponding storage value.

However, pySMT appends the bit width to the integer value in the
string representation of a bit vector, e.g., "1_256" in pySMT rather
than "1" in Z3. Therefore, calls to SymExecWrapper::find_storage_write()
which only use the stringified integer may fail, e.g., the ether_send
module and the external_calls module.

This commit fixes the above issue by adding the necessary suffix to
the `index` parameter of find_storage_write().