Fix the satisfaction check of jump condition
pySMT simplify() is somehow weaker than Z3 simplify() and cannot infer
that for all unsigned bit vector bv, bv < 0 does not hold. As a
consequence, the jump condition check in jumpi_() may fail and take
impossible paths.

To solve the above issue, we feed the jump condition to the constraint
solver to determine whether a path can be skipped.