Add a symbol manager
The symbol manager provides a unified and global symbol factory to
create all symbols. The symbol created by the symbol factory is
composed of
* the symbol object from pySMT (i.e., by pySMT API Symbol()), and
* additional properties attached to that symbol.

The properties are managed and used in Mythril, and are never used by
the constraint solvers. For example, we can store the taint flags in
symbol properties, as shown in `test_simple_taint_analysis()`.