analysis/integer: fix the overflow check of bit vector multiplication
Differently than the bit vector addition, an overflowed bit vector
multiplication does not necessarily result in a value less than any
of its operands (e.g., the multiplication of two 8-bit vectors 0x7 and
4 produces 0xC).

The previous attempt to fix this issue used Z3 API BVMulNoOverflow
which is only available in newer versions of Z3 and unfortunately
brought regressions. The fix in this commit implements a common
algorithm to check the overflow of symbolic bit vector multiplication
and does not reply on any new Z3 APIs.